testing 1 + some stuff
was testing some endpoints and the changes I made excluded some stuff.
Plus I had begun on cw2  before beginning this branch and I didn't send it to git so it keeps showing up as a changed file and its annoying to deal with for now. I will change it probably once I begin cw2.